ES6 contentScript - separate file for async loading reference: https://github.com/GanguLabs/chrome-extension-es6-import/blob/master/src/js/content_script.js
