+++ src/entries/contentScript/primary/main.js
@@ -1,4 +1,4 @@
-import highlightStack, { renderContent } from "../contentScript";
+import highlightStack, { renderContent } from "../contentScript_main.js";
 import App from "./App.svelte";
 
 // renderContent(import.meta.PLUGIN_WEB_EXT_CHUNK_CSS_PATHS, (appRoot) => {
